changed: separated test into mock.rs and tests.rs
